+++ package.json
@@ -12,6 +12,7 @@
     "@tanstack/react-query": "^5.83.0",
     "axios": "^1.10.0",
     "formik": "^2.4.6",
+    "jwt-decode": "^4.0.0",
     "next": "^15.3.5",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
