+++ src/context/AuthContext.tsx
@@ -1,85 +1,103 @@
-"use client";
+'use client';
 
-import {
-  createContext,
-  useContext,
-  useState,
-  ReactNode,
-  useEffect,
-} from "react";
-import IAuthContext from "../types/IAuthContext";
-import { User } from "../types/index";
+import { jwtDecode } from "jwt-decode";
+import { createContext, useContext, useEffect, useState } from "react";
 
-// Crear el contexto con valores iniciales vacios
+interface User {
+  email: string;
+  id_role: number;
+  name: string;
+  lastname: string;
+  phone: string;
+  identification_number: string;
+}
 
-const AuthContext = createContext<IAuthContext>({
-  user: null,
-  token: null,
-  login: () => {},
-  logout: () => {},
-  isLoggedIn: false,
-  setUser: () => {},
-});
+type SaveUserPayload = {
+  token: string;
+  login: boolean;
+}
 
-// Crear el provider que envuelve a la app y proveer el estado de autenticación
+type AuthContextType = {
+  //States
+  user?: User | null;
+  token?: string | null;
+  isAuth: boolean | null;
+  //Actions
+  saveUserData: (data: SaveUserPayload) => void;
+  resetUserData: () => void;
+};
 
-export const AuthProvider = ({ children }: { children: ReactNode }) => {
+const AuthContext = createContext<AuthContextType | undefined>(undefined);
+
+const USER_LOCAL_KEY = 'user';
+
+export const AuthProvider = ({children}: {children: React.ReactNode}) => {
   const [user, setUser] = useState<User | null>(null);
   const [token, setToken] = useState<string | null>(null);
+  const [isAuth, setIsAuth] = useState<AuthContextType['isAuth']>(null);
 
-  // Guardar usuario/token en localStorage cuando cambian
-  useEffect(() => {
-    if (user && token) {
-      localStorage.setItem("user", JSON.stringify(user));
-      localStorage.setItem("token", token);
-    } else {
-      localStorage.removeItem("user");
-      localStorage.removeItem("token");
+  const saveUserData = (data: SaveUserPayload) => {
+    setToken(data.token);
+    setIsAuth(data.login);
+
+    const payload = jwtDecode<User>(data.token)
+    console.log('El payload es:', payload);
+
+    const userData:User = {
+      email: payload.email,
+      id_role: payload.id_role,
+      name: payload.name,
+      lastname: payload.lastname,
+      phone: payload.phone,
+      identification_number: payload.identification_number,
     }
-  }, [user, token]);
-
-  // Recuperar usuario/token al cargar la app
-  useEffect(() => {
-    const storedUser = localStorage.getItem("user");
-    const storedToken = localStorage.getItem("token");
-    if (storedUser && storedToken) {
-      setUser(JSON.parse(storedUser));
-      setToken(storedToken);
-    }
-  }, []);
-
-  // creamos la funcion de login
-  const login = (userData: User, authToken: string) => {
+    
     setUser(userData);
-    setToken(authToken);
+    localStorage.setItem(USER_LOCAL_KEY, JSON.stringify({
+      token: data.token,
+      login: data.login,
+      user: userData,
+    }));
   };
 
-  // cramos la funcion de logout
-  const logout = () => {
+  const resetUserData = () => {
     setUser(null);
     setToken(null);
+    setIsAuth(false);
+    localStorage.removeItem(USER_LOCAL_KEY);
   };
 
-  // funcion que verifica si el usuario esta loggeado
-  const isLoggedIn = !!user && !!token;
+  useEffect(() => {
+    const storage = JSON.parse(localStorage.getItem(USER_LOCAL_KEY) || '{}');
 
-  // proveer los valores y funciones a los componentes hijos
+    if (storage === undefined || !Object.keys(storage)?.length) {
+      setIsAuth(false);
+      return;
+    };
+
+    const storageType: Storage = storage;
+    setUser(storage.user);
+    setToken(storageType?.token);
+    setIsAuth(storage?.login);
+  }, []);
 
   return (
-    <AuthContext.Provider
-      value={{
-        user,
-        token,
-        login,
-        logout,
-        isLoggedIn,
-        setUser,
-      }}
-    >
-      {children}
-    </AuthContext.Provider>
+    <AuthContext.Provider value={{
+      user,
+      token,
+      isAuth,
+      saveUserData,
+      resetUserData
+    }}>{children}</AuthContext.Provider>
   );
 };
 
-// hook para consumir el contexto en cualquier componente
-export const useAuth = () => useContext(AuthContext);
+export const useAuthContext = () => {
+  const context = useContext(AuthContext);
+
+  if (!context) {
+    throw new Error('El useAuthContext debe usarse dentro de un AuthProvider');
+  };
+
+  return context;
+};
